% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animated.hist.kernel.R
\name{animated.hist.kernel}
\alias{animated.hist.kernel}
\title{Animated histogram with varying kernel density estimates}
\usage{
animated.hist.kernel(x, nbreaks, bw.min, bw.max, bw.n = 20, outputname)
}
\arguments{
\item{x}{Variable of interest}

\item{nbreaks}{Number of breaks for histogram}

\item{bw.min}{Minimum bandwidth for kernel density}

\item{bw.max}{Maximum bandwidth for kernel density}

\item{bw.n}{Number of bandwidth numbers to consider for an animated plot. Default is 20.}

\item{outputname}{Output name to be saved}
}
\description{
Animated histogram with varying kernel density estimates
}
\examples{
setwd("/Users/seungjunahn/Documents/temp") #set the directory with a new folder
animated.hist.kernel(x=iris$Sepal.Length, nbreaks=30, bw.min=0.1, bw.max=0.5, bw.n=20, outputname="animated1")
}
