% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animated.monthly.report.R
\name{animated.monthly.report}
\alias{animated.monthly.report}
\title{Animated monthly plot}
\usage{
animated.monthly.report(data, start.date, end.date, months, outputname)
}
\arguments{
\item{data}{Input data}

\item{start.date}{Start date}

\item{end.date}{End date}

\item{months}{Months}

\item{outputname}{Name of the animated monthly plot}
}
\description{
Animated monthly plot
}
\examples{
animated.monthly.report(data=SanMartinoPPts, start.date="1970-01-01",
end.date="1989-12-31", months=1:12, outputname="animated2")
}
